requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.2
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
pymongo==4.6.2
python-dotenv==1.0.1
requests==2.31.0
schedule==1.2.1
tqdm==4.66.2
urllib3==2.2.1
Werkzeug==3.0.1
